@@ -29,8 +29,7 @@
     </dependencies>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>14</maven.compiler.source>
-        <maven.compiler.target>14</maven.compiler.target>
+            <maven.compiler.release>15</maven.compiler.release>
     </properties>
     <build>
         <finalName>frcdatawizard</finalName>
