@@ -480,11 +480,8 @@ public class StartHomeForm extends javax.swing.JFrame {
         //</editor-fold>
 
         /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new StartHomeForm().setVisible(true);
-
-            }
+        java.awt.EventQueue.invokeLater(() -> {
+            new StartHomeForm().setVisible(true);
         });
     }
 
