@@ -1,6 +1,5 @@
 package com.jr12221.frcdatawizard;
 
-
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
@@ -17,24 +16,24 @@ public class PreferenceReadWrite {
     int AddKeyNoInt(String user, String keyZ, boolean ftc) {
         try {
             try {
-                FileInputStream is = new FileInputStream(".frc-datawizard.properties");
-                propK.load(is);
-                is.close();
+                try (FileInputStream is = new FileInputStream(".frc-datawizard.properties")) {
+                    propK.load(is);
+                }
             } catch (FileNotFoundException e) {
                 System.out.println("File Not Found. Creating it.");
                 FileOutputStream fos = new FileOutputStream(".frc-datawizard.properties");
                 fos.close();
             }
-            FileOutputStream fos = new FileOutputStream(".frc-datawizard.properties");
-            if (ftc) {
-                propK.setProperty("usernameT", user);
-                propK.setProperty("keyT", keyZ);
-            } else {
-                propK.setProperty("usernameR", user);
-                propK.setProperty("keyR", keyZ);
+            try (FileOutputStream fos = new FileOutputStream(".frc-datawizard.properties")) {
+                if (ftc) {
+                    propK.setProperty("usernameT", user);
+                    propK.setProperty("keyT", keyZ);
+                } else {
+                    propK.setProperty("usernameR", user);
+                    propK.setProperty("keyR", keyZ);
+                }
+                propK.store(fos, "");
             }
-            propK.store(fos, "");
-            fos.close();
             return 0;
         } catch (IOException e) {
             System.out.println("I/O Error.  Oh no.  Check for write permissions.\n" + e.getMessage());
@@ -54,10 +53,10 @@ public class PreferenceReadWrite {
 
     void ListKey() {
         try {
-            FileInputStream is = new FileInputStream(".frc-datawizard.properties");
-            propK.load(is);
-            propK.forEach((k, v) -> System.out.println(k + "       " + v));
-            is.close();
+            try (FileInputStream is = new FileInputStream(".frc-datawizard.properties")) {
+                propK.load(is);
+                propK.forEach((k, v) -> System.out.println(k + "       " + v));
+            }
         } catch (IOException e) {
             System.out.println("I/O Error, maybe you haven't added your key with prefmgr -s? \n" + e.getMessage());
         }
@@ -67,9 +66,9 @@ public class PreferenceReadWrite {
         if (base3.equals(Main.FRC_BASE)) {
             try {
                 String encodedKey;
-                FileInputStream is = new FileInputStream(".frc-datawizard.properties");
-                propK.load(is);
-                is.close();
+                try (FileInputStream is = new FileInputStream(".frc-datawizard.properties")) {
+                    propK.load(is);
+                }
                 if (!propK.containsKey("usernameR") || !propK.containsKey("keyR")) {
                     System.out.println("Please enter your API key to use the program.");
                     return null;
@@ -85,9 +84,9 @@ public class PreferenceReadWrite {
         } else {
             try {
                 String encodedKey;
-                FileInputStream is = new FileInputStream(".frc-datawizard.properties");
-                propK.load(is);
-                is.close();
+                try (FileInputStream is = new FileInputStream(".frc-datawizard.properties")) {
+                    propK.load(is);
+                }
                 if (!propK.containsKey("usernameT") || !propK.containsKey("keyT")) {
                     System.out.println("Please enter your API key to use the program.");
                     return null;
@@ -106,18 +105,18 @@ public class PreferenceReadWrite {
     void setYear(int year) {
         try {
             try {
-                FileInputStream is = new FileInputStream(".frc-datawizard.properties");
-                propK.load(is);
-                is.close();
+                try (FileInputStream is = new FileInputStream(".frc-datawizard.properties")) {
+                    propK.load(is);
+                }
             } catch (FileNotFoundException e) {
                 System.out.println("File Not Found. Creating it.");
                 FileOutputStream fos = new FileOutputStream(".frc-datawizard.properties");
                 fos.close();
             }
             propK.setProperty("default_year", String.valueOf(year));
-            FileOutputStream fos = new FileOutputStream(".frc-datawizard.properties");
-            propK.store(fos, "");
-            fos.close();
+            try (FileOutputStream fos = new FileOutputStream(".frc-datawizard.properties")) {
+                propK.store(fos, "");
+            }
         } catch (IOException e) {
             System.out.println("I/O Error.  Oh no.  Check for write permissions.\n" + e.getMessage());
             System.exit(1);
@@ -126,9 +125,9 @@ public class PreferenceReadWrite {
 
     boolean ifYear() {
         try {
-            FileInputStream is = new FileInputStream(".frc-datawizard.properties");
-            propK.load(is);
-            is.close();
+            try (FileInputStream is = new FileInputStream(".frc-datawizard.properties")) {
+                propK.load(is);
+            }
             return propK.containsKey("default_year");
         } catch (IOException e) {
             return false;
@@ -138,9 +137,9 @@ public class PreferenceReadWrite {
 
     int getYear() {
         try {
-            FileInputStream is = new FileInputStream(".frc-datawizard.properties");
-            propK.load(is);
-            is.close();
+            try (FileInputStream is = new FileInputStream(".frc-datawizard.properties")) {
+                propK.load(is);
+            }
             return Integer.parseInt(propK.getProperty("default_year"));
         } catch (IOException e) {
             System.out.println("I/O Error.  This shouldn't happen.  Check for write permissions in $HOME. Exiting... \n" + e.getMessage());
@@ -151,9 +150,9 @@ public class PreferenceReadWrite {
 
     int getKeyCode() {
         try {
-            FileInputStream is = new FileInputStream(".frc-datawizard.properties");
-            propK.load(is);
-            is.close();
+            try (FileInputStream is = new FileInputStream(".frc-datawizard.properties")) {
+                propK.load(is);
+            }
         } catch (IOException e) {
             return -1;
         }
